6369-hmis-service-category-type-rules
